requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
pandas
openpyxl
python-dotenv
ipykernel
ddgs
jinja2
tqdm